Add new styles for About Us, Main Page, Rankings, and Ranking Selection pages
- Introduced AboutUs.css with a complete redesign inspired by Swedish design principles, including styles for banners, breadcrumbs, content sections, team sections, mission sections, contact forms, and responsive adjustments.
- Created MainPage.css for basic layout and video alignment.
- Developed Ranking.css with comprehensive styles for rankings page layout, headers, cards, tables, and responsive design.
- Added RankingSelection.css for landing page styles, including hero sections, ranking cards, and responsive adjustments for various screen sizes.
