+++ src/pages/MainPage/MainPage.js
@@ -1,8 +1,8 @@
-import "./MainPage.css"
-import News from '../News/News'
-import Ranking from '../Ranking/Ranking'
-import Videos from '../Videos/Videos'
-import IntroBanner from "../IntroBanner/IntroBanner"
+import "styles/MainPage.css"
+import News from '../../components/DynamicComponents/News/News'
+import Ranking from '../../components/DynamicComponents/Ranking/Ranking'
+import Videos from '../../components/DynamicComponents/Videos/Videos'
+import IntroBanner from "../../components/DynamicComponents/IntroBanner/IntroBanner"
 
 
 export default function MainPage(){
